Refactor project for use Clean Architecture with MVVM
Task:

- Reorganize the folder project
