+++ TVSeries/TVSeries/Infrastructure/Network/BaseError.swift
