+++ TVSeries/TVSeries/Infrastructure/Network/BaseMethod.swift
