+++ TVSeries/TVSeries/Presentation/Extensions/Encodable.swift
